requirements.txt
requests
psycopg2
PyYAML

